placewise-ai/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
pdfminer.six==20231228
spacy==3.7.4
scikit-learn==1.4.2
sentence-transformers==2.7.0
httpx==0.27.0
python-dotenv==1.0.1
numpy==1.26.4